Create cloudkit service account
This service account will be used by AAP in order to create the k8s
resources in the cluster.

We give the admin rights to this user for the moment in order to move
forward with the playbooks, but we will restraints its rights once we
figure out a better model.